shim: dock module
why not adapter?

Well, adapter kinda fits, but if you think about it, not quite.

What is adapter? First of all, the adapter came from sovereign. So it's
a term from the rollup development kit. However, on the shim side it
provides something else, specifically an infra supporting adapter
implementation, not the adapter implementation itself.

What it is, is kind of a port where the adapter could be inserted. A port
is a good name, but it's already has a connotation of "--port 80". To avoid
confusion, I picked "dock". A short concise name.