shim: fix submit key loading
This changeset fixes loading of the submit key. The problem was that if the key length was equal 32 then the function will return an error.

Also, I don't think it's a good idea to accept input that is not 32 bytes long, so I changes it to strict equallity check.

Test added.