shim: refactor the sugondat rpc client
This changeset makes a couple of fixes and changes:

1. Now the client strives to provide a peristent connection. See #68.
1. Fix the greedy polling (See #37). Now, only a single task polls for the
finalized header.
1. Extracts the timestamp from the extrinsics instead of getting it from the storage.
1. Also sprinkles more documentation, clarifies some assumptions.
1. Now checks whether the subxt codegen matches the runtime metadata. If not, then
we assume that the node is still useable, but still warn the user.
1. If the RPC node is not available, instead of bailing out, we
keep reconnecting. I imagine this should be useful in cases where we spin up
a local network and the shim started up before the sugondat node (plausible IMO).

Closes #68
Closes #56
Closes #37