build: docker-compose
This commit introduces docker/docker-compose.yml. Thanks to it
you can now spin up a network with just one command:

	docker compose -f docker/docker-compose.yml --build up

That will launch the polkadot localnet, the parachain node, the shim
and the demoes: gm (based on rollkit) and the sovereign demo.

Besides that this commit adds explicit cache IDs: that seems to improve
the caching behavior when running builds from docker-compose.

Atm, parity/polkadot image isn't multi-arch and doesn't support arm64.
To properly support macOS builds I opted to build the whole thing inside
amd64. It seems it has ok performance.

Also, I wrapped the docker entrypoints in tini. This solves the problem
of propagating the SIGTERM/SIGINT when stopping the services.