fees: add MaximumSkippedBlocks constant
Limit the number of skipped blocks in the fee update mechanism to the expected maximum.
This is done to avoid the possibility of incorrectly updating the fee multiplier.
For example, if the chain halts and skips more than 10000 blocks,
the approximation given by the first n terms of e^x could diverge significantly from the actual value.
Therefore, we set a boundary on the number of skipped blocks to ensure the accuracy of the approximation