update to asynchronous backing
Updating to asynchronous backing required changes to the runtime and the node.
It also involved changing the AdjustmentVariable to maintain a maximum variance
of 30% within one hour. This led to resolving some older todos that had been
left pending until the update to asynchronous backing was implemented,
which has been done differently than originally described (done keeping
the initial value for AdjustmentVariable).