shim: parallel submission
rollkit demo submits a lot of blobs. In my testing I saw
dozens of blobs submitted at the same time. Ofc submitting
those serially won't cut it, esp. so if waiting until
finalization.

This changeset solves the problem by splitting the signing
part, which is still serial, and the submission, which is
now parallel.

As a drive-by fix, all the errors for rollkit are now
encapsulated into an enum.

Closes #259