shim/rollkit: downgrade go-da to 0.2.0
the version that was introduced in #260
(997046d926170d0bcf5e2824d5670c0a5fb85a2f) was using
the latest version of go-da (0.4.0) gRPC service decl.

However, the latest version of rollkit on which the gm
demo was based off, was using 0.2.0. That was causing
occasional problems.

So I went ahead and downgraded it. Now, it seems to
workfine (with exception of #259).

Closes #258.